Refactor UserController and StudentRepository for async methods
Converted UserController and StudentRepository to use async methods for better performance and scalability. In the UserController, the UpdateUser method was also changed specifically to update an email, reflecting the required changes in StudentRepository and its interface. Password fields were added in UserDto to facilitate password change in the application. This modification allows the methods to be non-blocking and improves the application responsiveness giving a better user experience.