Change variable name to avoid confusion
The variable 'NewPassword' in UserDto.cs and 'newUser' in StudentRepository.cs was renamed to 'Password'. This was done to improve code readability and to ensure that other developers can understand the code more easily. The 'CurrentPassword' variable was also removed as it was not being used, making the code cleaner and more efficient.